+++ src/features/customers/CustomerRow.tsx
@@ -13,7 +13,7 @@ import { useDispatch } from "react-redux";
 import {
   defaultDeliveryDays,
   planLabels,
-  extrasLabels
+  extrasLabels,
 } from "../../lib/config";
 import deepMemo from "../../lib/deepMemo";
 import { getPlanString } from "../../lib/get-plan-string";
@@ -26,7 +26,7 @@ const SlimButton = styled(Button)`
   padding: 0 5px 0 5px;
 `;
 
-const UnMemoizedCustomerRow: React.FC<CustomerRowProps> = props => {
+const UnMemoizedCustomerRow: React.FC<CustomerRowProps> = (props) => {
   const [showDoDelete, setShowDoDelete] = React.useState(false);
   const [showPause, setShowPause] = React.useState(false);
   const [showEdit, setShowEdit] = React.useState(false);
@@ -43,7 +43,7 @@ const UnMemoizedCustomerRow: React.FC<CustomerRowProps> = props => {
       getPlanString(props.customer.newPlan, {
         planLabels: [...planLabels],
         extrasLabels: [...extrasLabels],
-        defaultDeliveryDays: [...defaultDeliveryDays]
+        defaultDeliveryDays: [...defaultDeliveryDays],
       }),
     [props.customer]
   );
@@ -59,7 +59,7 @@ const UnMemoizedCustomerRow: React.FC<CustomerRowProps> = props => {
       <TableCell>
         {props.customer.exclusions.length > 0
           ? props.customer.exclusions
-              .map(exclusion => exclusion.name)
+              .map((exclusion) => exclusion.name)
               .join(", ")
           : "None"}
       </TableCell>
@@ -107,7 +107,7 @@ const UnMemoizedCustomerRow: React.FC<CustomerRowProps> = props => {
               const customer = {
                 ...props.customer,
                 pauseStart: undefined,
-                pauseEnd: undefined
+                pauseEnd: undefined,
               };
               dispatch(updateCustomer(customer));
             }}