+++ src/meal-planning/create-variant.ts
@@ -7,6 +7,22 @@ import { SelectedItem, SelectedMeal } from "./types";
 const hasExclusions = (exclusion: Exclusion, meal: Recipe | undefined) =>
   meal?.potentialExclusions.some((value) => value.id === exclusion.id);
 
+const createExclusionsString = (
+  customer: Customer,
+  item: SelectedMeal,
+  allMeals: Recipe[]
+) => {
+  const realMeal = allMeals.find((theMeal) => theMeal.id === item.recipe.id);
+
+  const matchingExclusions = customer.exclusions.filter((allergen) => {
+    return realMeal?.potentialExclusions.some(
+      (value) => value.id === allergen.id
+    );
+  });
+
+  return matchingExclusions.map((exclusion) => exclusion.name).join(", ");
+};
+
 const createVariantString = (
   customer: Customer,
   item: SelectedItem,
@@ -16,19 +32,11 @@ const createVariantString = (
     return item.chosenVariant;
   }
 
-  const realMeal = allMeals.find((theMeal) => theMeal.id === item.recipe.id);
+  const exclusions = createExclusionsString(customer, item, allMeals);
 
-  const matchingExclusions = customer.exclusions.filter((allergen) => {
-    return realMeal?.potentialExclusions.some(
-      (value) => value.id === allergen.id
-    );
-  });
+  const exclusionsString = exclusions ? `$ (${exclusions})` : ``;
 
-  return matchingExclusions.length > 0
-    ? `${item.chosenVariant} (${matchingExclusions
-        .map((exclusion) => exclusion.name)
-        .join(", ")})`
-    : `${item.chosenVariant}`;
+  return `${item.chosenVariant}${exclusionsString}`;
 };
 
 const createMealWithVariantString = (
@@ -46,6 +54,7 @@ export const createVariant = (
   customisation: boolean;
   allergen: boolean;
   string: string;
+  exclusions?: string;
   mealWithVariantString: string;
 } => {
   if (!isSelectedMeal(meal)) {
@@ -73,6 +82,7 @@ export const createVariant = (
   return {
     customisation: matchingExclusions.length > 0,
     allergen: matchingExclusions.length > 0 && matchingExclusions[0].allergen,
+    exclusions: createExclusionsString(customer, meal, allMeals),
     mealWithVariantString: createMealWithVariantString(
       customer,
       meal,