+++ src/scripts/clone-prod-db.ts
@@ -13,25 +13,25 @@ const doCopy = async (sourceEnv: string, destinationEnv: string) => {
       "exclusions-table",
       "customers-table",
       "customer-exclusions-table",
-      "recipe-exclusions-table"
+      "recipe-exclusions-table",
     ].map(
-      async table =>
+      async (table) =>
         await new Promise((resolve, reject) => {
           copy(
             {
               config: {
                 accessKeyId: process.env.AWS_ACCESS_KEY_ID,
                 secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
-                region: "us-east-1"
+                region: "us-east-1",
               },
               source: {
-                tableName: `${sourceEnv}-tnm-admin-${table}`
+                tableName: `${sourceEnv}-tnm-admin-${table}`,
               },
               destination: {
-                tableName: `${destinationEnv}-tnm-admin-${table}`
+                tableName: `${destinationEnv}-tnm-admin-${table}`,
               },
               log: true,
-              create: false
+              create: false,
             },
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
             (error: Error | null, result: any) => {